# docs/swagger.yaml
basePath: /api
definitions:
...
  handler.changeStatusRequest:
    properties:
      status:
        description: |-
          int 而非 domainuser.Status：0 是合法取值（禁用），required 会拒绝 0；
          用 min/max 限定 0~1，swag 也能正确生成 schema
        maximum: 1
        minimum: 0
        type: integer
      userId:
        type: integer
    required:
    - userId
    type: object
...
  handler.loginRequest:
    properties:
      ip:
        type: string
      password:
        type: string
      username:
        type: string
    required:
    - password
    - username
    type: object
...
  handler.registerRequest:
    properties:
      password:
        maxLength: 20
        minLength: 6
        type: string
      phone:
        type: string
      username:
        maxLength: 10
        minLength: 2
        type: string
    required:
    - password
    - phone
    - username
    type: object
...
  handler.updateUserRequest:
    properties:
      avatar:
        type: string
      birthday:
        type: string
      email:
        type: string
      gender:
        maximum: 2
        minimum: 0
        type: integer
      nickname:
        maxLength: 10
        minLength: 2
        type: string
      phone:
        type: string
    required:
    - nickname
    type: object
...
  response.Response:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
    type: object
...
  user.LoginChannel:
    enum:
    - 0
    - 1
    - 2
    type: integer
    x-enum-comments:
      ChannelAlipay: 支付宝快捷登录
      ChannelPassword: 账号密码登录
      ChannelWechat: 微信登录(小程序/公众号/App)
    x-enum-descriptions:
    - 账号密码登录
    - 微信登录(小程序/公众号/App)
    - 支付宝快捷登录
    x-enum-varnames:
    - ChannelPassword
    - ChannelWechat
    - ChannelAlipay
  user.Status:
    enum:
    - 0
    type: integer
    x-enum-varnames:
    - StatusDisabled
  userapp.GetUserListResult:
    properties:
      list:
        items:
          $ref: '#/definitions/userapp.UserDTO'
        type: array
      total:
        type: integer
    type: object
  userapp.LoginResult:
    properties:
      token:
        type: string
      user:
        $ref: '#/definitions/userapp.UserDTO'
    type: object
  userapp.UserDTO:
    properties:
      alipayUid:
        type: string
      avatar:
        type: string
      birthday:
        type: string
      createdAt:
        type: string
      email:
        type: string
      gender:
        type: integer
      lastLoginAt:
        type: string
      lastLoginIp:
        type: string
      loginChannel:
        $ref: '#/definitions/user.LoginChannel'
      nickname:
        type: string
      openId:
        type: string
      phone:
        type: string
      status:
        $ref: '#/definitions/user.Status'
      unionId:
        type: string
      updatedAt:
        type: string
      userId:
        type: integer
      username:
        type: string
    type: object
host: localhost:9000
info:
  contact: {}
  description: DDD 架构示例服务（用户 / 商品目录 / 订单）。host 为网关端口，经网关按前缀转发到各服务
  title: Demo Service API
  version: "1.0"
paths:
...
  /user/list:
    get:
      description: 分页查询用户，支持按用户ID精确过滤；需要登录（JWT）
      parameters:
      - description: 页码（从1开始）
        in: query
        name: page
        required: true
        type: integer
      - description: 每页数量（1~100）
        in: query
        name: pageSize
        required: true
        type: integer
      - description: 用户ID（精确匹配）
        in: query
        name: userId
        type: integer
      - description: 用户名
        in: query
        name: username
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/userapp.GetUserListResult'
              type: object
        "400":
          description: 参数校验失败
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: 未登录或 token 失效
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - ApiKeyAuth: []
      summary: 分页查询用户列表
      tags:
      - 用户
  /user/login:
    post:
      consumes:
      - application/json
      description: 登录成功返回 JWT token 与用户信息；用户名或密码错误返回 401（不区分账号是否存在），账号被禁用返回 403
      parameters:
      - description: 登录信息
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.loginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/userapp.LoginResult'
              type: object
        "400":
          description: 参数校验失败
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: 用户名或密码错误
          schema:
            $ref: '#/definitions/response.Response'
        "403":
          description: 账号已被禁用
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/response.Response'
      summary: 用户名密码登录
      tags:
      - 用户
  /user/register:
    post:
      consumes:
      - application/json
      description: 用户名密码注册。用户名已存在时返回 400
      parameters:
      - description: 注册信息
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.registerRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/userapp.UserDTO'
              type: object
        "400":
          description: 参数校验失败或用户名已存在
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/response.Response'
      summary: 注册用户
      tags:
      - 用户
  /user/status:
    put:
      consumes:
      - application/json
      description: 启用或禁用指定用户。status：0-禁用，1-正常。需要登录（JWT）
      parameters:
      - description: 用户ID与目标状态
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.changeStatusRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: 参数校验失败
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: 未登录或 token 失效
          schema:
            $ref: '#/definitions/response.Response'
        "404":
          description: 用户不存在
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - ApiKeyAuth: []
      summary: 修改用户状态
      tags:
      - 用户
  /user/user/profile:
    put:
      consumes:
      - application/json
      description: 更新 JWT 对应用户的基础资料，用户ID取自 token。整体替换语义：资料字段以本次提交为准，未提交的可选字段（phone/email/avatar/gender/birthday）会被置空。需要登录（JWT）
      parameters:
      - description: 用户资料
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.updateUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/userapp.UserDTO'
              type: object
        "400":
          description: 参数校验失败或性别取值不合法
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: 未登录或 token 失效
          schema:
            $ref: '#/definitions/response.Response'
        "404":
          description: 用户不存在
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - ApiKeyAuth: []
      summary: 更新当前用户资料
      tags:
      - 用户
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"